backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-smtp-ssl==1.0
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
twilio==7.16.0
tzdata==2022.7
urllib3==1.26.13
